:heavy_plus_sign: Create main.py and setup.py
As of now setup.py is empty. It will contain code that will setup the file system when the user uses the password manager for the first time.

Inside main.py, code has been written to display the main menu to the user. Two options are displayed

1. Sign up and create new vault
2. Login and view vault

The users response is then recorded using the functionality provided by rich.Prompt